requirements.txt
google-cloud-bigquery>=3.20.0
psycopg2-binary>=2.9.0
schedule>=1.2.0
